sentence_similarity/backend/requirements.txt
click==7.1.2
databases==0.3.2
fastapi==0.58.0
uvicorn==0.11.5
idna==2.9
importlib-metadata==1.6.1
motor==2.1.0
nltk==3.5
numpy==1.19.0
packaging==20.4
pydantic==1.5.1
pymongo==3.10.1
pyparsing==2.4.7
python-dotenv==0.13.0
python-slugify==4.0.0
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5
sentence-transformers==0.2.6.1
six==1.15.0
spacy==2.3.0
SQLAlchemy==1.3.18
starlette==0.13.4
tokenizers==0.7.0
zipp==3.1.0